+++ packages/editor/src/index.ts
@@ -154,6 +154,7 @@ export {
 	type PointsSnapIndicator,
 	type SnapIndicator,
 } from './lib/editor/managers/SnapManager/SnapManager'
+export { Session } from './lib/editor/sessions/Session'
 export { BaseBoxShapeUtil, type TLBaseBoxShape } from './lib/editor/shapes/BaseBoxShapeUtil'
 export {
 	ShapeUtil,
