+++ packages/editor/src/lib/components/default-components/DefaultCanvas.tsx
@@ -367,7 +367,6 @@ function SelectedIdIndicators() {
 				editor.isInAny(
 					'select.idle',
 					'select.brushing',
-					'select.scribble_brushing',
 					'select.editing_shape',
 					'select.pointing_shape',
 					'select.pointing_selection',
