+++ packages/tldraw/src/test/translating.test.ts
@@ -137,7 +137,7 @@ describe('When translating...', () => {
 		jest.advanceTimersByTime(100)
 		editor
 			// The change is bigger than expected because the camera moves
-			.expectShapeToMatch({ id: ids.box1, x: -160, y: 10 })
+			.expectShapeToMatch({ id: ids.box1, x: -180, y: 10 })
 			// We'll continue moving in the x postion, but now we'll also move in the y position.
 			// The speed in the y position is smaller since we are further away from the edge.
 			.pointerMove(0, 25)
@@ -155,13 +155,12 @@ describe('When translating...', () => {
 	it('translates a single shape near the bottom right edge', () => {
 		editor.user.updateUserPreferences({ edgeScrollSpeed: 1 })
 		editor.pointerDown(50, 50, ids.box1).pointerMove(1080, 50)
-
-		jest.advanceTimersByTime(100)
+		editor.forceTick(4)
 		editor
 			// The change is bigger than expected because the camera moves
 			.expectShapeToMatch({ id: ids.box1, x: 1140, y: 10 })
 			.pointerMove(1080, 800)
-		jest.advanceTimersByTime(100)
+		editor.forceTick(6)
 		editor
 			.expectShapeToMatch({ id: ids.box1, x: 1280, y: 845.68 })
 			.pointerUp()
